chore: Refactor shim builder
